implemented refactoring: api and html should be extracted #17 https://github.com/nobio/timetrack/issues/17
